# Performance optimization of CurrentLineVerifier