--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -3,4 +3,4 @@
 ## 2.1.0
 - added IonProperty ref-struct that represents property row of IIonReader
 - added extension method IIonReader.ReadProperty() that returns IonProperty struct
-
+- ION reading performance optimization