before:
```
| Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|------:|------:|------:|----------:|
|   Read | 6.077 ms | 0.0869 ms | 0.0813 ms |     - |     - |     - |   7.92 KB |
```

after:
```
| Method |     Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|----------:|----------:|------:|------:|------:|----------:|
|   Read | 5.722 ms | 0.0571 ms | 0.0534 ms |     - |     - |     - |   7.92 KB |
```